Organize code based on reviewer's suggestions
