+++ src/nvim/os/path_win.h
@@ -1,0 +1,7 @@
+#ifndef NVIM_OS_PATH_WIN_H
+#define NVIM_OS_PATH_WIN_H
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "os/path_win.h.generated.h"
+#endif
+#endif  // NVIM_OS_PATH_WIN_H
